fix: Добавил ScoreParser Выделил всю логику, разбирающую входные данные в этот парсер
